backend/pyproject.toml
[project]
name = "health-assistant-backend"
version = "0.1.0"
description = "건강상담도우미 FastAPI 백엔드"
requires-python = ">=3.10"
dependencies = [
    # LangChain + LangGraph (v1 LTS)
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langgraph>=1.0,<2.0",
    "langchain-google-genai",
    # Observability
    "langsmith>=0.3.0",
    # Web Framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sse-starlette>=2.0.0",
    # Database
    "supabase>=2.4.0",
    "asyncpg>=0.29.0",
    # HTTP Client
    "httpx>=0.27.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # Auth (Clerk JWT)
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    # PDF Generation
    "reportlab>=4.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
